issue-107-create-statics-data-class - 그래프 데이터 클래스 요일 타입 수정

+++ app/src/main/java/teamh/boostcamp/myapplication/data/model/Statics.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public class Statics {
 
-    private int week; // 요일
+    private float week; // 요일
     private float emotion; // 감정
     private List<Entry> weekEmotionData; // 요일과 감정으로 만든 하나의 좌표 데이터
 
     // 생성자를 통한 주입
-    public Statics(int week, float emotion) {
+    public Statics(float week, float emotion) {
         weekEmotionData = new ArrayList<>();
         this.week = week;
         this.emotion = emotion;
